BasicSelect retirado de Adapters; CitiesService.listByState adicionado
