+++ frontend/src/views/components/Adapters.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react'
 import { TextField, Label, DatePicker } from 'office-ui-fabric-react'
-import Select from 'react-select'
 import VirtualizedSelect from 'react-virtualized-select'
 import { Field } from 'react-final-form'
 
 import CitySelect from './CitySelect'
+import BasicSelect from './BasicSelect'
 
 export const TextFieldAdapter = ({ input, label, meta, ...rest }) => (
   <TextField
@@ -82,26 +82,3 @@ export const CitySelectAdapter = ({ input, label, options, ...rest }) => (
     onChange={value => input.onChange(value)}
   />
 )
-
-class BasicSelect extends Component {
-  constructor (props) {
-    super(props)
-
-    this.state = {
-      value: undefined
-    }
-  }
-
-  onChange (input, name) {
-    this.setState({ value: input.value })
-    this.props.onChange(input.value)
-  }
-
-  render () {
-    return <Select
-      {...this.props}
-      value={this.state.value}
-      onChange={value => this.onChange(value, this.props.name)}
-    />
-  }
-}
